Push tenant subscriptions to the package level
